LPS-70036 These methods are only used in unit tests
